requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Deprecated==1.2.14
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.28.1
pycparser==2.21
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
wrapt==1.15.0
yarl==1.9.2
